+++ assets/javascripts/kitten/components/atoms/loader-with-paragraph/stories.js
@@ -1,18 +1,13 @@
 import React from 'react'
 import { text, select } from '@storybook/addon-knobs'
-import { Loader } from './loader'
-import { LoaderWithParagraph } from './loader-with-paragraph'
+import { LoaderWithParagraph } from './index'
 
 export default {
-  title: 'Loaders/Loaders',
-  component: Loader,
+  title: 'Atoms/LoaderWithParagraph',
+  component: LoaderWithParagraph,
 }
 
 export const Default = () => {
-  return <Loader color={text('Color', 'black')} />
-}
-
-export const WithParagraph = () => {
   return (
     <LoaderWithParagraph
       loaderPosition={select(
@@ -21,7 +16,7 @@ export const WithParagraph = () => {
         'left',
       )}
     >
-      {text('Wording', 'Loading')}
+      {text('children', 'Loading')}
     </LoaderWithParagraph>
   )
 }